Adding fixed widescreens
including unity assets and statusbar stuff

+++ src/hexen/r_segs.c
@@ -18,7 +18,6 @@
 #include "h2def.h"
 #include "i_system.h"
 #include "r_local.h"
-#include "dpplimits.h"
 
 // OPTIMIZE: closed two sided lines as single sided
 
@@ -349,7 +348,7 @@ void R_StoreWallRange(int start, int stop)
 // calculate rw_distance for scale calculation
 //
     rw_normalangle = curline->angle + ANG90;
-    offsetangle = abs((int) rw_normalangle - (int) rw_angle1);
+    offsetangle = abs(rw_normalangle - rw_angle1);
     if (offsetangle > ANG90)
         offsetangle = ANG90;
     distangle = ANG90 - offsetangle;
